Simplify further by creating a dummy object
This dummy object never gets moved and therefore has the original controller position that represent the position of the physical controller.

This is much, much simpler and more reliable and more accurate than doing any math calculations to get this position and then check if it is inside some collider or not....

The dummy object can just use the OnColliderExit event.